Daydreamer with DreamerV3

+++ embodied/core/path.py
@@ -110,6 +110,10 @@ class Path:
   def copy(self, dest):
     raise NotImplementedError
 
+  def move(self, dest):
+    self.copy(dest)
+    self.remove()
+
 
 class LocalPath(Path):
 
@@ -147,7 +151,13 @@ class LocalPath(Path):
     shutil.rmtree(self)
 
   def copy(self, dest):
-    shutil.copytree(self, type(self)(dest), dirs_exist_ok=True)
+    if self.isfile():
+      shutil.copy(self, type(self)(dest))
+    else:
+      shutil.copytree(self, type(self)(dest), dirs_exist_ok=True)
+
+  def move(self, dest):
+    shutil.move(self, dest)
 
 
 class GFilePath(Path):
@@ -198,6 +208,12 @@ class GFilePath(Path):
 
   def copy(self, dest):
     self._gfile.copy(str(self), str(dest), overwrite=True)
+
+  def move(self, dest):
+    dest = Path(dest)
+    if dest.isdir():
+      dest.rmtree()
+    self._gfile.rename(self, str(dest), overwrite=True)
 
 
 Path.filesystems = [
